pyproject.toml
[tool.poetry]
name = "lada-discord-bot"
version = "0.1.0"
description = ""
authors = ["urielmiralles <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.7.1"
discord-py-slash-command = "^1.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
